requirements.txt
google-adk
python-dotenv
google-generativeai
reportlab
streamlit
openpyxl
# sounddevice 
# soundfile
# openai-whisper